Rename Button and Input components
Temporary.
