+++ compass/app/page.tsx
@@ -1,8 +1,8 @@
 // pages/index.tsx
 "use client";
 
-import Button from '@/components/Button';
-import Input from '@/components/Input'
+import Button from '@/components/Button49';
+import Input from '@/components/Input49'
 import InlineLink from '@/components/InlineLink';
 import Paper from '@/components/auth/Paper';
 import Image from 'next/image';
